Move changes to a subclass or CueChunk
